FIX: Reduce issues when debouncing schema search
Previously the debounced value was both setting the filter value and
updating the input's text value. This causes visible UI issues, because
the debounce updating the input's text value would sometimes reset it,
especially if/when typing quickly.

This separates the two values: `filter` is still used to handle the
table filtering, but `filterInput` is used as the input value, and it is
not tracked.

This PR also uses sticky positioning for the search form, so that it
stays visible even when scrolling the list of schema tables.