FIX: Query downloads were being passed an incorrect query object. (#359)
This is a follow-up to d726c4889e5c5d7726ed1035c37d0cd60a3bb3d6.

The previous change missed changing the name of the query object when passing it to QueryResultsWrapper, which resulted in the download links not working properly after a query was run.

This change fixes that bug, and includes an acceptance test to ensure it stays fixed during future work on this plugin.