FIX: Discourse automation reports with parameters (#363)
When setting up an automation to create a DM or a post with a report on a recurring basis, and using a Data Explorer query that has parameters, we encounter an error.

The `.params_to_hash` method currently expects an array of arrays for the parameters, but in reality they seem to be arrays of hashes.

This change makes `.params_to_hash` work with arrays of hashes. It also preserves the ability to work with nested arrays in case this is used somewhere else.