FIX: Ensure dates passed in query params are parsed in the local timezone. (#365)
When passing date values for reports as URL parameters, they were being parsed by the standard JavaScript Date object, which due to legacy reasons parses date-only strings in UTC, all other date formats are assumed to be in the browser's local timezone.

This change switches to using moment's parser, which assumes that all strings passed to it are in the browser's local timezone.